+++ invenio_rdm_records/__init__.py
@@ -10,6 +10,6 @@
 
 from .ext import InvenioRDMRecords
 
-__version__ = "4.13.1"
+__version__ = "4.15.1"
 
 __all__ = ("__version__", "InvenioRDMRecords")