+++ invenio_rdm_records/services/communities/components.py
@@ -7,6 +7,8 @@
 
 """Record communities service components."""
 
+from flask import current_app
+from invenio_access.permissions import system_identity
 from invenio_communities.communities.records.systemfields.access import VisibilityEnum
 from invenio_communities.communities.services.components import (
     CommunityAccessComponent as BaseAccessComponent,
@@ -18,11 +20,13 @@ from invenio_communities.communities.services.components import (
     OwnershipComponent,
     PIDComponent,
 )
+from invenio_drafts_resources.services.records.components import ServiceComponent
 from invenio_i18n import lazy_gettext as _
 from invenio_records_resources.services.records.components import (
     MetadataComponent,
     RelationsComponent,
 )
+from invenio_requests.tasks import request_moderation
 from invenio_search.engine import dsl
 
 from ...proxies import current_community_records_service
@@ -57,6 +61,25 @@ class CommunityAccessComponent(BaseAccessComponent):
             self._check_visibility(identity, record)
 
 
+class ContentModerationComponent(ServiceComponent):
+    """Service component for content moderation."""
+
+    def create(self, identity, data=None, record=None, **kwargs):
+        """Create a moderation request if the user is not verified."""
+        if current_app.config["RDM_USER_MODERATION_ENABLED"]:
+            # If the publisher is the system process, we don't want to create a moderation request.
+            # Even if the record being published is owned by a user that is not system
+            if identity == system_identity:
+                return
+
+            # resolve current user and check if they are verified
+            is_verified = identity.user.verified_at is not None
+
+            if not is_verified:
+                # Spawn a task to request moderation.
+                request_moderation.delay(identity.id)
+
+
 CommunityServiceComponents = [
     MetadataComponent,
     CustomFieldsComponent,
@@ -66,4 +89,5 @@ CommunityServiceComponents = [
     OwnershipComponent,
     FeaturedCommunityComponent,
     OAISetComponent,
+    ContentModerationComponent,
 ]