+++ invenio_rdm_records/services/permissions.py
@@ -18,6 +18,7 @@ from invenio_records_permissions.generators import (
     SystemProcess,
 )
 from invenio_records_permissions.policies.records import RecordPermissionPolicy
+from invenio_requests.services.generators import Creator, Receiver, Status
 from invenio_requests.services.permissions import (
     PermissionPolicy as RequestPermissionPolicy,
 )
@@ -26,6 +27,7 @@ from ..requests.access import GuestAccessRequest
 from .generators import (
     AccessGrant,
     GuestAccessRequestToken,
+    IfCreate,
     IfExternalDOIRecord,
     IfFileIsLocal,
     IfNewRecord,
@@ -185,7 +187,7 @@ class RDMRecordPermissionPolicy(RecordPermissionPolicy):
     # Record communities
     #
     # Who can add record to a community
-    can_add_community = [RecordOwners(), SystemProcess()]
+    can_add_community = can_manage
     # Who can remove a community from a record
     can_remove_community = [
         RecordOwners(),
@@ -196,7 +198,7 @@ class RDMRecordPermissionPolicy(RecordPermissionPolicy):
     can_remove_record = [CommunityCurators()]
 
     #
-    # Media files
+    # Media files - draft
     #
     can_draft_media_create_files = can_review
     can_draft_media_read_files = can_review
@@ -215,6 +217,24 @@ class RDMRecordPermissionPolicy(RecordPermissionPolicy):
     can_draft_media_delete_files = can_review
 
     #
+    # Media files - record
+    #
+    can_media_read_files = [
+        IfRestricted("record", then_=can_view, else_=can_all),
+        ResourceAccessToken("read"),
+    ]
+    can_media_get_content_files = [
+        # note: even though this is closer to business logic than permissions,
+        # it was simpler and less coupling to implement this as permission check
+        IfFileIsLocal(then_=can_read, else_=[SystemProcess()])
+    ]
+    can_media_create__files = [Disable()]
+    can_media_set_content_files = [Disable()]
+    can_media_commit_files = [Disable()]
+    can_media_update_files = [Disable()]
+    can_media_delete_files = [Disable()]
+
+    #
     # Disabled actions (these should not be used or changed)
     #
     # - Records/files are updated/deleted via drafts so we don't support
@@ -226,6 +246,10 @@ class RDMRecordPermissionPolicy(RecordPermissionPolicy):
     can_commit_files = [Disable()]
     can_update_files = [Disable()]
     can_delete_files = [Disable()]
+
+    # Used to hide at the moment the `parent.is_verified` field. It should be set to
+    # correct permissions based on which the field will be exposed only to moderators
+    can_moderate = [Disable()]
 
 
 guest_token = IfRequestType(
@@ -240,3 +264,20 @@ class RDMRequestsPermissionPolicy(RequestPermissionPolicy):
     can_update = RequestPermissionPolicy.can_update + [guest_token]
     can_action_submit = RequestPermissionPolicy.can_action_submit + [guest_token]
     can_action_cancel = RequestPermissionPolicy.can_action_cancel + [guest_token]
+    can_create_comment = can_read
+    can_update_comment = RequestPermissionPolicy.can_update_comment + [guest_token]
+    can_delete_comment = RequestPermissionPolicy.can_delete_comment + [guest_token]
+
+    # manages GuessAccessRequest payload permissions
+    can_manage_access_options = [
+        IfCreate(
+            then_=[SystemProcess()],
+            else_=[
+                IfRequestType(
+                    GuestAccessRequest,
+                    then_=[Status(["submitted"], [Receiver()])],
+                    else_=Disable(),
+                )
+            ],
+        )
+    ]