+++ invenio_rdm_records/services/tasks.py
@@ -10,6 +10,7 @@
 from celery import shared_task
 from flask import current_app
 from invenio_access.permissions import system_identity
+from invenio_i18n import lazy_gettext as _
 
 from invenio_rdm_records.proxies import current_rdm_records
 from invenio_rdm_records.services.errors import EmbargoNotLiftedError