+++ tests/requests/test_user_moderation_actions.py
@@ -5,37 +5,59 @@
 # # Invenio-RDM is free software; you can redistribute it and/or modify
 # # it under the terms of the MIT License; see LICENSE file for more details.
 """Test user moderation actions."""
-from time import sleep
-
+from celery import Task
 from invenio_access.permissions import system_identity
-from invenio_requests.proxies import current_requests_service
+from invenio_db import db
+from invenio_requests.proxies import (
+    current_requests_service,
+    current_user_moderation_service,
+)
 
 from invenio_rdm_records.proxies import current_rdm_records_service as records_service
 
 
+class MockRequestModerationTask(Task):
+    """Mock celery task for moderation request."""
+
+    def delay(*args):
+        user_id = args[0]
+        with db.session.begin_nested():
+            try:
+                current_user_moderation_service.request_moderation(
+                    system_identity, user_id=user_id, uow=None
+                )
+            except Exception as ex:
+                pass
+
+
 def test_user_moderation_approve(
-    running_app,
-    mod_request_create,
-    mod_identity,
-    uploader,
-    es_clear,
-    minimal_record,
+    running_app, mod_identity, unverified_user, es_clear, minimal_record, mocker
 ):
     """Tests moderation action after user approval.
 
     The user records, and all its versions, should have a "is_verified" field set to "True".
     """
 
-    user_records_q = f"parent.access.owned_by.user:{uploader.id}"
+    user_records_q = f"parent.access.owned_by.user:{unverified_user.id}"
 
     # Create a record
-    draft = records_service.create(uploader.identity, minimal_record)
-    record = records_service.publish(id_=draft.id, identity=uploader.identity)
+    draft = records_service.create(unverified_user.identity, minimal_record)
+    record = records_service.publish(id_=draft.id, identity=unverified_user.identity)
 
     # Create a new version of the record
-    new_version = records_service.new_version(uploader.identity, id_=record.id)
-    records_service.update_draft(uploader.identity, new_version.id, minimal_record)
-    new_record = records_service.publish(identity=uploader.identity, id_=new_version.id)
+    new_version = records_service.new_version(unverified_user.identity, id_=record.id)
+    records_service.update_draft(
+        unverified_user.identity, new_version.id, minimal_record
+    )
+    # Since tasks are executed synchronously in tests, the db session used for publishing is rolledback.
+    # This is a patch for tests only.
+    mocker.patch(
+        "invenio_rdm_records.services.components.verified.request_moderation",
+        MockRequestModerationTask,
+    )
+    new_record = records_service.publish(
+        identity=unverified_user.identity, id_=new_version.id
+    )
 
     pre_approval_records = records_service.search(
         system_identity, params={"q": user_records_q, "allversions": True}
@@ -47,11 +69,17 @@ def test_user_moderation_approve(
 
     assert is_verified == False
 
-    moderation_request = mod_request_create(uploader.id)
+    # Fetch moderation request that was created on publish
+    res = current_requests_service.search(
+        system_identity, params={"q": f"topic.user:{unverified_user.id}"}
+    )
+    assert res.total == 1
+
+    mod_request = res.to_dict()["hits"]["hits"][0]
 
     # Approve user, records are verified from now on
     current_requests_service.execute_action(
-        mod_identity, id_=moderation_request.id, action="accept"
+        mod_identity, id_=mod_request["id"], action="accept"
     )
 
     # Records are re-indexed in bulk. We need to force it for tests, otherwise it's done by a celery beat task